Add help messages to clean.sh and show_kb.sh

+++ clean.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ $1 = "-h" || $1 = "--help" ]]; then
+    printf "\nUsage: ./clean.sh - Removes ALL docker containers and all <none> or 'ostis' taged images\n\n"
+    exit 0
+fi
+
 echo "Cleaning..."
 
 kill -9 $(pgrep "run.sh") > /dev/null 2>&1
